+++ NEWS.md
@@ -7,7 +7,8 @@ data sets for Nature paper and SBS1-SBS5 paper fail to reproduce the data sets
 used by these papers.
 
 In SynSigGen 1.1.0, we disabled the rounding of mutations due to each signature
-when running these data generators, so that these legacy data sets can be reproduced.
+when running generator functions for legacy data sets, so that these legacy data
+sets can be reproduced.
 
 
 ## SynSigGen 1.0.13